viral-engine/requirements.txt
fastapi
uvicorn
yt-dlp
moviepy==1.0.3
groq
openai-whisper
python-multipart
pydantic
dotenv